Moved stringify function to kafka/sarama
