+++ pkg/channel/distributed/common/util/util.go
@@ -20,8 +20,6 @@ import (
 	"os"
 	"os/signal"
 
-	"github.com/Shopify/sarama"
-
 	"go.uber.org/zap"
 )
 
@@ -37,24 +35,4 @@ func WaitForSignal(logger *zap.Logger, signals ...os.Signal) {
 
 	// Log Signal Receipt
 	logger.Info("Received Signal", zap.String("Signal", sig.String()))
-}
-
-// Utility function to convert []byte headers to string ones for logging purposes
-func StringifyHeaders(headers []sarama.RecordHeader) map[string][]string {
-	stringHeaders := make(map[string][]string)
-	for _, header := range headers {
-		key := string(header.Key)
-		stringHeaders[key] = append(stringHeaders[key], string(header.Value))
-	}
-	return stringHeaders
-}
-
-// Pointer-version of the StringifyHeaders function
-func StringifyHeaderPtrs(headers []*sarama.RecordHeader) map[string][]string {
-	stringHeaders := make(map[string][]string)
-	for _, header := range headers {
-		key := string(header.Key)
-		stringHeaders[key] = append(stringHeaders[key], string(header.Value))
-	}
-	return stringHeaders
 }
